requirements.txt
asgiref==3.5.2
Django==3.2.5
django-filter==22.1
djangorestframework==3.12
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
